+++ src/views/small-program/components/collect.vue
@@ -1,0 +1,11 @@
+<!--
+ * @Author: 刘一笔
+ * @Date: 2023-05-02 21:58:42
+ * @Description: 我收藏的
+-->
+<template>
+  <div>2</div>
+</template>
+
+<script setup>
+</script>
